Move formatters into common 'formatters' directory
