+++ src/libraries/formatters/duration-formatter.js
