+++ src/libraries/formatters/time-formatter.js
