+++ src/main/dependencies/modules/bug-reporting.js
@@ -27,7 +27,7 @@ import LogCache from '../../../app/logging/log-cache'
 import LogCacheBundle from '../../../app/logging/log-cache-bundle'
 import type { BugReporterMetrics } from '../../../app/bug-reporting/metrics/bug-reporter-metrics'
 import BugReporterMetricsStore from '../../../app/bug-reporting/metrics/bug-reporter-metrics-store'
-import { TimeFormatter } from '../../../libraries/time-formatter'
+import { TimeFormatter } from '../../../libraries/formatters/time-formatter'
 
 function bootstrap (container: Container) {
   container.factory(
