+++ test/unit/app/bug-reporting/metrics/bug-reporter-metrics-store.spec.js
@@ -20,7 +20,7 @@
 import { beforeEach, describe, expect, it } from '../../../../helpers/dependencies'
 import BugReporterMetricsStore from '../../../../../src/app/bug-reporting/metrics/bug-reporter-metrics-store'
 import { EXTRA, METRICS, NOT_SET, TAGS } from '../../../../../src/app/bug-reporting/metrics/metrics'
-import { TimeFormatter } from '../../../../../src/libraries/time-formatter'
+import { TimeFormatter } from '../../../../../src/libraries/formatters/time-formatter'
 
 describe('BugReporterMetricsStore', () => {
   let store: BugReporterMetricsStore
